Create a function that take a location
Use fetch with the url of an api plus with the location
Return json file so they can be use for the app.

Create a function that can take a location and weatherapi within
the object, and second object made for elements to replace the
default value of html elements with the return values of the weather
api.